Added Docteur Controller's wonderful theme.  This theme also helped me find an issue in the theme updater logic. I will do some touchups to this theme before we reach the next release version, but that will take a while due to the major code changes in progress.

+++ ThemeFixer/OSD_Fixer.cs
@@ -43,9 +43,9 @@ namespace ThemeFixer
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public float? rot { get; set; }
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int? width { get; set; }
+        public float? width { get; set; }
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int? height { get; set; }
+        public float? height { get; set; }
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public bool? center { get; set; }
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
